requirements.txt
Django
gunicorn
djangorestframework
pandas
scikit-learn
openpyxl
psycopg[binary]
django-cors-headers
django-environ  # for environment variables
redis
django-redis
kaleido
pyarrow
lxml
defusedxml
djangorestframework-simplejwt
dj-rest-auth
django-allauth
requests
cryptography
pyotp
user-agents
Pillow
reportlab